✨ Refactor widget structure in MainApp
Reorganized MainApp widget structure for better readability and clarity. Including moving MainApp widget definition to a separate file for separation of concerns.
